// src/models/IoK8sApiNetworkingV1IngressClassSpec/types.ts
import { IoK8sApiNetworkingV1IngressClassParametersReferenceModelOptions as IoK8sApiNetworkingV1IngressClassParametersReference } from '../IoK8sApiNetworkingV1IngressClassParametersReference';

export interface IoK8sApiNetworkingV1IngressClassSpecModelOptions {
  controller?: string;
  parameters?: IoK8sApiNetworkingV1IngressClassParametersReference;
}